requirements.txt
fastapi
uvicorn
pymupdf
requests
python-dotenv
jose
passlib
sqlalchemy
langchain
langchain-cohere
langchain-community
chromadb